added new link actions and behaviour
More optimized link behaviour. Unified linkAction function that is
called. When ctrl is pressed a new tab is created. When shift is pressed
the popup will be rewritten if possible, else replace the current tab.

Also replaced focus on the serach input with instant click.

Added vim key binding. "j" for down "k" for up